orchestrator/requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
PyJWT==2.8.0
requests==2.31.0
twilio==8.10.1
urllib3<2
yarl==1.9.2
python-dotenv==1.0.0
